Migrate Boolean Couter to single class (#1366)
Removed the three classes (BooleanCounter, BooleanCouterController,
BooleanCounterParameters) used to represent the boolean counter
processing element and replaced them with a single class,
BooleanCounterProcessor.
